pyproject.toml
[tool.poetry]
name = "sie"
version = "0.1.0"
description = "A simple image editor"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Pillow = "^8.2.0"
PyQt5 = "^5.15.4"
psd-tools = "^1.9.18"

[tool.poetry.scripts]
sie = "sie.main:main"